refactored bouts endpoint to a thin aggregator and exported bouts api functionality to api/bouts_routes
