+++ backend/app/api/bouts_routes/__init__.py
@@ -1,0 +1,1 @@
+"""Composable route modules for the /bouts API surface."""
